Refactor: 예외 처리 관련 로직 수정
